+++ resources/js/components/ui/alert-dialog/AlertDialogDescription.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import { AlertDialogDescription } from 'radix-vue'
+import { cn } from '@/lib/utils'
+
+defineProps<{
+  class?: string
+}>()
+</script>
+
+<template>
+  <AlertDialogDescription :class="cn('text-sm text-muted-foreground', $props.class)">
+    <slot />
+  </AlertDialogDescription>
+</template>