+++ src/Http/Controllers/ClearMailboxController.php
@@ -2,17 +2,22 @@
 
 namespace Redberry\MailboxForLaravel\Http\Controllers;
 
-use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\View\View;
 use Redberry\MailboxForLaravel\CaptureService;
 
 class ClearMailboxController
 {
-    public function __invoke(Request $request, CaptureService $service): View|JsonResponse
+    public function __invoke(Request $request, CaptureService $service): RedirectResponse
     {
         $service->clearAll();
 
-        return response()->json();
+        return redirect()->back()->with([
+            'flash' => [
+                'status' => 'success',
+                'title' => 'Mailbox cleared.',
+                'description' => 'All messages have been successfully deleted.',
+            ],
+        ]);
     }
 }